rework of filter
